Resolve merge conflicts
Update all references to `src/view`, which has been renamed to
`src/components`.
