+++ theme/src/components/ProductImageCarousel/carousel.css
