+++ theme/src/components/ProductImageCarousel/carousel.js
