+++ theme/src/components/ProductImageCarousel/thumbnail.css
