+++ theme/src/components/ProductImageCarousel/thumbnail.js
