+++ theme/src/components/ProductImageCarousel/thumbnailList.css
