+++ theme/src/components/ProductOptions/swatch.js
