+++ theme/src/components/ProductQuantity/quantity.css
