+++ theme/src/components/RichText/index.js
