+++ theme/src/components/RichText/richText.js
