+++ theme/src/components/Select/option.js
